Refactor: create useSubMenuState custom hook
